xin shi

+++ Furnish/MainWindow.xaml.cs
@@ -14,6 +14,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using System.Windows.Shell;
 
 
 namespace Furnish
@@ -33,6 +34,8 @@ namespace Furnish
                 MessageBox.Show(this, "Unable to access the database:\n" + ex.Message, "Fatal database error", MessageBoxButton.OK, MessageBoxImage.Error);
                 Environment.Exit(1);
             }
+
+            
         }
 
     private void BtnOrders_Click(object sender, RoutedEventArgs e)
@@ -51,21 +54,68 @@ namespace Furnish
             dialog.Owner = this;
             dialog.ShowDialog();
         }
-    private void BtnLogin_Click(object sender, RoutedEventArgs e)
-    {
 
-        Login dialog = new Login(); 
-        dialog.Owner = this;
-            
-        if (dialog.ShowDialog() == true)
-        {
-            LvUsers.ItemsSource = Globals.dbContext.Employees.ToList(); 
-            TbkStatus.Text = "Employee login";
+        static Employee Emp = null;
+
+        public static List<Employee> GetList()
+        {   
+            List<Employee> Emplist = new List<Employee>();
+
+            string _name = App.Current.Properties[1].ToString();
+     
+            Emp = (from employee in Globals.dbContext.Employees where employee.name == _name select employee).FirstOrDefault<Employee>();
+
+            if(Emp == null)
+            {
+                Emplist= null;
+            }
+            else if (Emp.role == RoleEnum.User)
+            {
+                Emplist.Add(Emp);
+
+            }
+            else
+            {
+                Emplist = Globals.dbContext.Employees.ToList();
+            }
+
+            return Emplist;
         }
-    }
+
+
+
+        private void BtnLogin_Click(object sender, RoutedEventArgs e)
+    {
+            
+            if (Btxlogin.Text == "Login")
+            {
+                Login dialog = new Login();
+                dialog.Owner = this;
+                if (dialog.ShowDialog() == true)
+                {
+                    Btxlogin.Text = "Logout";
+                    
+                    LvUsers.ItemsSource = GetList();
+                    TbkStatus.Text = "welcome: " + App.Current.Properties[1].ToString();
+                }
+            }
+            else if(Btxlogin.Text == "Logout")
+            {
+                Emp = null;
+                App.Current.Properties[0] ="";
+
+                App.Current.Properties[1] = "";
+                App.Current.Properties[2] = "";
+                App.Current.Properties[3] = "";
+                LvUsers.ItemsSource = GetList(); 
+                Btxlogin.Text = "Login";
+                TbkStatus.Text = "";
+            }
+        }
 
     private void BtnProducts_Click(object sender, RoutedEventArgs e)
     {
+
         ProductsViewDlg dialog = new ProductsViewDlg();
         dialog.Owner = this;
         dialog.ShowDialog();
@@ -91,7 +141,7 @@ namespace Furnish
 
             if (dialog.ShowDialog() == true)
             {
-                LvUsers.ItemsSource = Globals.dbContext.Employees.ToList();
+                LvUsers.ItemsSource = GetList();
                 TbkStatus.Text = "Employee updated";
             }
             else
@@ -112,12 +162,15 @@ namespace Furnish
 
         private void MenuItem_ListDeleteItemClick(object sender, RoutedEventArgs e)
         {
-
-
             Employee currEmp = LvUsers.SelectedItem as Employee;
-            Globals.dbContext.Employees.Remove(currEmp);            
-            Globals.dbContext.SaveChanges();
-            TbkStatus.Text = "Employee deleted";
+            if (currEmp.name != (string)App.Current.Properties[1]) {
+                Globals.dbContext.Employees.Remove(currEmp);
+                Globals.dbContext.SaveChanges();
+                LvUsers.ItemsSource = GetList();
+                TbkStatus.Text = "Employee deleted";
+            }
+            
+            
         }
